+++ src/pages/Lend/LoanChannelCategoryControl.vue
@@ -76,13 +76,13 @@
 
 					<helpme-choose-wrapper
 						v-if="showHelpMeChooseFeat"
-						:remaining-loans="helpmeChooseRemainingLoans"
+						:remaining-loans="helpMeChooseRemainingLoans"
 						:items-in-basket="itemsInBasket"
 						:is-visitor="isVisitor"
 						:user-data="userData"
 						:loan-channel-name="loanChannelName"
-						:loans="helpmeChooseLoans"
-						@update="getHelpmeChooseLoans($event)"
+						:loans="helpMeChooseLoans"
+						@update="getHelpMeChooseLoans($event)"
 						:is-loading="isLoadingHC"
 					/>
 				</div>
@@ -138,13 +138,13 @@
 
 					<helpme-choose-wrapper
 						v-if="showHelpMeChooseFeat"
-						:remaining-loans="helpmeChooseRemainingLoans"
+						:remaining-loans="helpMeChooseRemainingLoans"
 						:items-in-basket="itemsInBasket"
 						:is-visitor="isVisitor"
 						:user-data="userData"
 						:loan-channel-name="loanChannelName"
-						:loans="helpmeChooseLoans"
-						@update="getHelpmeChooseLoans($event)"
+						:loans="helpMeChooseLoans"
+						@update="getHelpMeChooseLoans($event)"
 						:is-loading="isLoadingHC"
 						:enable-loan-tags="enableLoanTags"
 						:enable-loan-card-exp="enableLoanCardExp"
@@ -191,7 +191,8 @@ import {
 	getCachedChannel,
 	watchChannelQuery,
 	getLoanChannel,
-	getMetaDescription
+	getMetaDescription,
+	getFLSSQueryMap,
 } from '@/util/loanChannelUtils';
 import { runFacetsQueries, fetchLoanFacets } from '@/util/loanSearch/dataUtils';
 import { transformIsoCodes } from '@/util/loanSearch/filters/regions';
@@ -311,10 +312,6 @@ export default {
 		};
 	},
 	props: {
-		enableHelpmeChoose: {
-			type: Boolean,
-			default: false,
-		},
 		enableLoanTags: {
 			type: Boolean,
 			default: false
@@ -375,7 +372,8 @@ export default {
 			selectedQuickFilters: {},
 			userData: {},
 			showQuickFiltersOverlay: false,
-			helpmeChooseLoans: [],
+			helpMeChooseSort: null,
+			helpMeChooseLoans: [],
 			isLoadingHC: true,
 		};
 	},
@@ -413,7 +411,7 @@ export default {
 			}
 			return _filter(this.allLoans, (loan, index) => index > 0);
 		},
-		helpmeChooseRemainingLoans() {
+		helpMeChooseRemainingLoans() {
 			if (this.displayLoanPromoCard) {
 				return _filter(this.allLoans, (loan, index) => index > 4);
 			}
@@ -463,7 +461,12 @@ export default {
 			return this.totalCount <= this.limit;
 		},
 		showHelpMeChooseFeat() {
-			return this.enableHelpmeChoose && this.allLoans.length > 8;
+			const queryMapFLSS = getFLSSQueryMap(this.loanChannelQueryMap, this.targetedLoanChannelURL);
+			const hasSortBy = !!queryMapFLSS.sortBy;
+
+			// Don't show help me choose if the category has sortBy
+			// Help me choose categories are just different sortBy options
+			return !hasSortBy && this.allLoans.length > 8;
 		}
 	},
 	apollo: {
@@ -624,6 +627,7 @@ export default {
 			}
 			this.activateLoanChannelWatchQuery();
 			this.resetPagination();
+			this.getHelpMeChooseLoans();
 		},
 		checkIfPageIsOutOfRange(loansArrayLength, pageQueryParam) {
 			// determines if the page query param is for a page that is out of bounds.
@@ -803,8 +807,13 @@ export default {
 				this.flssLoanSearch = matchedUrls[0]?.flssLoanSearch ?? {};
 			}
 		},
-		async getHelpmeChooseLoans() {
+		async getHelpMeChooseLoans(sortBy) {
 			this.isLoadingHC = true;
+
+			if (sortBy) {
+				this.helpMeChooseSort = sortBy;
+			}
+
 			const loansData = await getLoanChannel(
 				this.apollo,
 				this.loanChannelQueryMap,
@@ -814,10 +823,13 @@ export default {
 					limit: 3,
 					basketId: this.cookieStore.get('kvbskt'),
 					origin: FLSS_ORIGIN_CATEGORY
-				}
+				},
+				// Apply quick filters to help me choose
+				{ ...this.selectedQuickFilters, sortBy: this.helpMeChooseSort }
 			);
+
 			const loans = loansData?.lend?.loanChannelsById[0]?.loans?.values ?? [];
-			this.helpmeChooseLoans = loans;
+			this.helpMeChooseLoans = loans;
 			this.isLoadingHC = false;
 		},
 		resetPagination() {